Fixes type for 'product_type' remove old indexing tests
